Fix clang warning
(cherry picked from commit 9e5cd1f79a5d384e4d7ab8d6169f1f35a6fb3d2c)

+++ src/gui/raster/qgspalettedrendererwidget.cpp
@@ -863,7 +863,7 @@ void QgsPalettedRendererClassGatherer::run()
         }
       }
       i ++;
-      emit progressChanged( 100 * ( i / static_cast<float>( newClasses.count() ) ) );
+      emit progressChanged( 100 * ( static_cast< double >( i ) / static_cast<double>( newClasses.count() ) ) );
     }
     mClasses = newClasses;
   }
